refactor: add small styling changes
