+++ src/widgets/carousel/slide/index.tsx
